Move PlateMapExcelWriter to from api to assay package
